--- io-actions/jest.setup.js
@@ -1,8 +1,0 @@
-/*
- * <license header>
- */
-
-jest.setTimeout(10000);
-
-beforeEach(() => {});
-afterEach(() => {});
