+++ io-actions/package.json
@@ -13,11 +13,13 @@
   "devDependencies": {
     "@openwhisk/wskdebug": "^1.3.0",
     "eslint": "^8",
-    "eslint-plugin-jest": "^27.2.3",
-    "jest": "^29.7.0"
+    "mocha": "^10.2.0",
+    "chai": "^4.3.6",
+    "esm": "^3.2.25",
+    "nock": "^13.0.0"
   },
   "scripts": {
-    "test": "jest --passWithNoTests ./test",
+    "test": "mocha --recursive ./test",
     "lint": "eslint --ignore-pattern web-src --no-error-on-unmatched-pattern test src actions",
     "lint:fix": "npm run lint -- --fix"
   },
