+++ io-actions/src/fragment/translate.js
@@ -6,16 +6,19 @@ const PATH_TOKENS =
 /**
  * we expect a body to already have been fetched, and a locale to be requested
  */
-async function main({ status, body, locale }) {
+async function main({ status, message, body, locale }) {
     if (!status || status != 200 || !body || !locale) {
         return {
             status: 400,
+            message:
+                message || 'requested source is either not here or invalid',
         };
     }
     const match = body.path?.match(PATH_TOKENS);
     if (!match) {
         return {
             status: 400,
+            message: 'source path is either not here or invalid',
         };
     }
     const { surface, parsedLocale, fragmentPath } = match.groups;
@@ -23,15 +26,16 @@ async function main({ status, body, locale }) {
         const response = await fetch(
             `https://odin.adobe.com/adobe/sites/fragments?path=/content/dam/mas/${surface}/${locale}/${fragmentPath}`,
         );
-        const resp = await response.json();
-        if (resp?.items?.length == 1) {
+        const root = await response.json();
+        if (root?.items?.length == 1) {
             return {
                 status: 200,
-                body: resp.items[0],
+                body: root.items[0],
             };
         } else {
             return {
-                status: 400,
+                status: 404,
+                message: 'no translation found',
             };
         }
     }
