+++ io-actions/test/setup.js
@@ -1,0 +1,16 @@
+const nock = require('nock');
+
+beforeEach(() => {
+    nock.cleanAll();
+    nock.disableNetConnect();
+});
+
+afterEach(() => {
+    nock.enableNetConnect();
+});
+
+module.exports = {
+    get fetchStub() {
+        return fetchStub;
+    },
+};
